feat: add JobsTasksForEachTaskConfigs for improved task configuration and validation
